+++ Jmol/src/org/openscience/jmol/viewer/managers/StyleManager.java
@@ -112,15 +112,15 @@ public class StyleManager {
   }
 
   // FIXME NEEDSWORK -- arrow vector stuff
-  public double arrowHeadSize = 10.0f;
-  public double arrowHeadRadius = 1.0f;
-  public double arrowLengthScale = 1.0f;
+  public float arrowHeadSize = 10.0f;
+  public float arrowHeadRadius = 1.0f;
+  public float arrowLengthScale = 1.0f;
 
-  public void setArrowHeadSize(double ls) {
+  public void setArrowHeadSize(float ls) {
     arrowHeadSize = 10.0f * ls;
   }
 
-  public double getArrowHeadSize() {
+  public float getArrowHeadSize() {
     return arrowHeadSize / 10.0f;
   }
 
@@ -128,23 +128,23 @@ public class StyleManager {
   // for some reason, internal to ArrowLine the raw arrowHeadSize was
   // used, but externally it is multiplied/divided by 10
   // will figure it out and fix it later
-  public double getArrowHeadSize10() {
+  public float getArrowHeadSize10() {
     return arrowHeadSize;
   }
 
-  public void setArrowLengthScale(double ls) {
+  public void setArrowLengthScale(float ls) {
     arrowLengthScale = ls;
   }
 
-  public double getArrowLengthScale() {
+  public float getArrowLengthScale() {
     return arrowLengthScale;
   }
 
-  public void setArrowHeadRadius(double rs) {
+  public void setArrowHeadRadius(float rs) {
     arrowHeadRadius = rs;
   }
 
-  public double getArrowHeadRadius() {
+  public float getArrowHeadRadius() {
     return arrowHeadRadius;
   }
 }
